+++ tests/testthat/test-oc_check_status.R
@@ -1,10 +1,15 @@
 ## Test oc_check_status ##
 
-# keys from https://opencagedata.com/api#codes or
-# https://github.com/OpenCageData/opencagedata-misc-docs/blob/master/library-guidelines.md # nolint
-key_402 <- "4372eff77b8343cebfc843eb4da4ddc4" # always returns a 402 responce
-key_403 <- "2e10e5e828262eb243ec0b54681d699a" # always returns a 403 responce
-key_429 <- "d6d0f0065f4348a4bdfe4587ba02714b" # always returns a 429 responce
+test_that("oc_check_status returns no error if HTTP status 200", {
+  skip_on_cran()
+  skip_if_offline()
+
+  withr::local_envvar(c("OPENCAGE_KEY" = key_200))
+  expect_type(
+    oc_reverse(latitude = 0, longitude = 0),
+    "list"
+  )
+})
 
 test_that("oc_check_status returns 400 error if request is invalid", {
   skip_on_cran()
@@ -12,11 +17,21 @@ test_that("oc_check_status returns 400 error if request is invalid", {
 
   # Both shouldn't happen since we oc_check_query
   expect_error(
-    oc_process(latitude = 280, longitude = 0, return = "json_list"),
+    oc_process(
+      latitude = 280,
+      longitude = 0,
+      key = Sys.getenv("OPENCAGE_KEY"),
+      return = "json_list"
+    ),
     "HTTP failure: 400"
   )
+
   expect_error(
-    oc_process(placename = "", return = "json_list"),
+    oc_process(
+      placename = "",
+      key = Sys.getenv("OPENCAGE_KEY"),
+      return = "json_list"
+    ),
     "HTTP failure: 400"
   )
 })
@@ -25,8 +40,9 @@ test_that("oc_check_status returns 401 error if key is invalid", {
   skip_on_cran()
   skip_if_offline()
 
+  withr::local_envvar(c("OPENCAGE_KEY" = "32charactersandnumbers1234567890"))
   expect_error(
-    oc_reverse(latitude = 0, longitude = 0, key = "thisisaninvalidkey"),
+    oc_reverse(latitude = 0, longitude = 0),
     "HTTP failure: 401"
   )
 })
@@ -35,8 +51,9 @@ test_that("oc_check_status returns 402 error if quota exceeded", {
   skip_on_cran()
   skip_if_offline()
 
+  withr::local_envvar(c("OPENCAGE_KEY" = key_402))
   expect_error(
-    oc_reverse(latitude = 0, longitude = 0, key = key_402),
+    oc_reverse(latitude = 0, longitude = 0),
     "HTTP failure: 402"
   )
 })
@@ -45,8 +62,9 @@ test_that("oc_check_status returns 403 error if key is blocked", {
   skip_on_cran()
   skip_if_offline()
 
+  withr::local_envvar(c("OPENCAGE_KEY" = key_403))
   expect_error(
-    oc_reverse(latitude = 0, longitude = 0, key = key_403),
+    oc_reverse(latitude = 0, longitude = 0),
     "HTTP failure: 403"
   )
 })
@@ -55,8 +73,9 @@ test_that("oc_check_status returns 429 error if rate limit is exceeded", {
   skip_on_cran()
   skip_if_offline()
 
+  withr::local_envvar(c("OPENCAGE_KEY" = key_429))
   expect_error(
-    oc_reverse(latitude = 0, longitude = 0, key = key_429),
+    oc_reverse(latitude = 0, longitude = 0),
     "HTTP failure: 429"
   )
 })